Update description, setup guidelines for OpenAPI and create auction component
This commit updates the README file with the project description, guidelines on how to generate client code using OpenAPI tools, and explanation for same. It also introduces a new auction component which displays whisky auction details in a table format. Notably, we've added a predefined command in the package.json file to support OpenAPI client code generation. Alongside, new styles for component display have been added in the global styles file. We are utilizing an OpenAPI specification file (whisky-service.yaml) for code generation. It's vital to remember updating this file as required before initiating a code generation process.